[NFC][lldb] move DiagnosticsRendering to Host
