+++ lldb/source/Host/CMakeLists.txt
@@ -17,6 +17,7 @@ macro(add_host_subdirectory group)
 endmacro()
 
 add_host_subdirectory(common
+  common/DiagnosticsRendering.cpp
   common/FileAction.cpp
   common/FileCache.cpp
   common/File.cpp
